포트폴리오

+++ GameEngineCore/GameEngineTileMap.h
@@ -47,7 +47,7 @@ public:
 
     GameEngineCollision* GetTileCollision(int _ZIndex, float4 _Pos); 
 
-    bool IsValidIndex(int _Z, int _Y, int _X);
+    bool IsValidIndex(int _Z, float _Y, float _X);
 
     float4 GetIndex(float4 _Pos);
